+++ src/features/auth/screens/Login/LoginScreen.jsx
@@ -1,46 +1,27 @@
 // src/features/auth/screens/LoginScreen.jsx
 import React, {useState} from "react";
-import {
-  View,
-  Text,
-  StyleSheet,
-  TextInput,
-  TouchableOpacity,
-  KeyboardAvoidingView,
-  Platform,
-  TouchableWithoutFeedback,
-  Keyboard,
-} from "react-native";
-import {
-  login,
-  getProfile,
-  logout,
-  unlink,
-} from "@react-native-seoul/kakao-login";
-
-// 🔽 SVG 아이콘 import (경로는 프로젝트에 맞게 수정)
-import BetaLogo from "@shared/assets/svg/logos/BetaLogo.svg";
-import KakaoIcon from "./assets/kakao.svg";
-import NaverIcon from "./assets/naver.svg";
+import {View, Text, StyleSheet, TouchableOpacity, Alert} from "react-native";
 import {SafeAreaView} from "react-native-safe-area-context";
-import {kakaoSignIn} from "./libs/kakaoSignIn";
-import {naverSignIn} from "./libs/naverSignIn";
-import {useSocialLoginMutation} from "../../services/socialLoginMutation";
+import {login, unlink} from "@react-native-seoul/kakao-login";
+
 import AuthBackground from "../../components/AuthBackground";
-import {useSignupSecretStore} from "../../stores/useSignupSecretStore";
+import {kakaoSignIn} from "../../libs/Login/kakaoSignIn";
+import {naverSignIn} from "../../libs/Login/naverSignIn";
+import {useSocialLoginMutation} from "../../services/socialLoginMutation";
+
+// 아이콘(svg) - 프로젝트 경로에 맞게 유지
+import BetaLogo from "@shared/assets/svg/logos/BetaLogo.svg";
+import KakaoIcon from "../../assets/Login/kakao.svg";
+import NaverIcon from "../../assets/Login/naver.svg";
+import AppleIcon from "../../assets/Login/apple.svg";
 
 const LoginScreen = ({navigation}) => {
-  const {setSocialAuth, clearSecrets} = useSignupSecretStore();
-
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
   const [isSocialLoading, setIsSocialLoading] = useState(false);
-
   const socialLoginMutation = useSocialLoginMutation();
 
-  const handleLogin = () => {
-    console.log("login", {email, password});
-    navigation.navigate("Main");
+  const handleAppleLogin = () => {
+    // TODO: 애플 로그인 붙일 때 여기 구현
+    Alert.alert("준비 중", "애플 로그인은 아직 연결되지 않았어요.");
   };
 
   const handleKakaoLogin = async () => {
@@ -49,35 +30,58 @@ const LoginScreen = ({navigation}) => {
 
     try {
       const {token, profile, cancelled} = await kakaoSignIn();
-
       if (cancelled) return;
 
-      // 여기서부터는 너의 앱 로직 👇
       console.log("카카오 토큰:", token);
       console.log("카카오 프로필:", profile);
 
-      // 1) 백엔드에 토큰 보내서 우리 앱용 accessToken 발급
-      // const { appToken } = await kakaoSignInAndIssueAppToken();
+      // ✅ 카카오 버튼 눌렀을 때 기존 로직 그대로 동작
       socialLoginMutation.mutate(
         {provider: "KAKAO", token: token.accessToken},
         {
           onSuccess: (response) => {
-            console.log("소셜 로그인 성공! newUser?:", response.data.newUser);
-          },
-        },
+            console.log("소셜 로그인 성공! newUser?:", response?.data?.newUser);
 
-        {
+            // 필요하면 여기서 이동 처리
+            // navigation.replace("Main");
+          },
           onError: (error) => {
             console.log("소셜 로그인 실패:", error);
+            Alert.alert("카카오 로그인 실패", "잠시 후 다시 시도해주세요.");
           },
-        }
+        },
       );
-      // 2) Zustand / AsyncStorage / MMKV 등에 appToken 저장
-      // clearSecrets(); // 혹시 이전 가입 시도 흔적 제거
-      // setSocialAuth({provider: "KAKAO", token: token.accessToken});
-      // 3) 홈 화면으로 이동 (navigation.navigate("Home") 같은 것)
     } catch (error) {
       Alert.alert("카카오 로그인 실패", "잠시 후 다시 시도해주세요.");
+    } finally {
+      setIsSocialLoading(false);
+    }
+  };
+
+  const handleNaverLogin = async () => {
+    if (isSocialLoading) return;
+    setIsSocialLoading(true);
+
+    try {
+      const {token, profile, cancelled} = await naverSignIn();
+      if (cancelled) return;
+
+      console.log("네이버 토큰:", token);
+      console.log("네이버 프로필:", profile);
+
+      // ✅ 필요하면 카카오처럼 백엔드 소셜 로그인 붙일 수 있음
+      // socialLoginMutation.mutate(
+      //   {provider: "NAVER", token: token.accessToken},
+      //   {
+      //     onSuccess: (response) => navigation.replace("Main"),
+      //     onError: () => Alert.alert("네이버 로그인 실패", "잠시 후 다시 시도해주세요."),
+      //   }
+      // );
+
+      // 지금은 기존 로직(로그/확인)만 유지
+    } catch (error) {
+      console.log(error);
+      Alert.alert("네이버 로그인 실패", "잠시 후 다시 시도해주세요.");
     } finally {
       setIsSocialLoading(false);
     }
@@ -97,136 +101,59 @@ const LoginScreen = ({navigation}) => {
     }
   };
 
-  const handleNaverLogin = async () => {
-    if (isSocialLoading) return;
-    setIsSocialLoading(true);
-
-    try {
-      const {token, profile, cancelled} = await naverSignIn();
-
-      if (cancelled) return;
-
-      console.log("네이버 토큰:", token); // accessToken, refreshToken, 만료시간 등
-      console.log("네이버 프로필:", profile); // email, name, nickname, gender, age 등
-
-      // 3) 홈 화면으로 이동 (navigation.navigate("Home") 같은 것)
-
-      // 1) 백엔드에 네이버 토큰 보내서 우리 앱용 accessToken 발급
-      socialLoginMutation.mutate(
-        {provider: "KAKAO", token: token.accessToken},
-        {
-          onSuccess: (response) => {
-            console.log("소셜 로그인 성공! newUser?:", response.data.newUser);
-          },
-        },
-
-        {
-          onError: (error) => {
-            console.log("소셜 로그인 실패:", error);
-          },
-        }
-      );
-      // clearSecrets(); // 혹시 이전 가입 시도 흔적 제거
-      // setSocialAuth({provider: "NAVER", token: token.accessToken});
-      // 2) MMKV / AsyncStorage / Zustand 등에 appToken 저장
-      // 3) 메인 화면 이동
-
-      // navigation.replace("Main");
-    } catch (error) {
-      console.log(error);
-      Alert.alert("네이버 로그인 실패", "잠시 후 다시 시도해주세요.");
-    } finally {
-      setIsSocialLoading(false);
-    }
-  };
-
   return (
     <SafeAreaView style={styles.safeArea}>
-      <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-        <KeyboardAvoidingView
-          style={styles.inner}
-          // behavior={Platform.OS === "ios" ? "padding" : undefined}
-        >
-          <AuthBackground />
+      <View style={styles.container}>
+        <AuthBackground />
 
-          {/* BETA 로고 텍스트 */}
-          <View style={styles.logoWrapper}>
-            <BetaLogo width={120} />
-            {/* <Text style={styles.logoText}>BETA</Text> */}
-          </View>
+        {/* 중앙 로고 텍스트 */}
+        <View style={styles.logoArea}>
+          <BetaLogo width={150} height={50} />
+        </View>
 
-          {/* 이메일 / 비밀번호 입력 */}
-          <View style={styles.form}>
-            <TextInput
-              style={styles.input}
-              placeholder="이메일을 입력하세요"
-              placeholderTextColor="#B8B8C4"
-              value={email}
-              onChangeText={setEmail}
-              keyboardType="email-address"
-              autoCapitalize="none"
-            />
-            <TextInput
-              style={[styles.input, styles.inputSpacing]}
-              placeholder="비밀번호를 입력하세요"
-              placeholderTextColor="#B8B8C4"
-              value={password}
-              onChangeText={setPassword}
-              secureTextEntry
-            />
+        {/* 하단 버튼 영역 */}
+        <View style={styles.bottomArea}>
+          <TouchableOpacity
+            style={[styles.fullButton, styles.appleButton]}
+            onPress={handleAppleLogin}
+            activeOpacity={0.85}
+            disabled={isSocialLoading}
+          >
+            <AppleIcon width={20} height={20} />
+            <Text style={[styles.fullButtonText, styles.appleText]}>
+              Apple 로그인
+            </Text>
+          </TouchableOpacity>
 
-            {/* 로그인 버튼 */}
-            <TouchableOpacity style={styles.loginButton} onPress={handleLogin}>
-              <Text style={styles.loginButtonText}>로그인</Text>
-            </TouchableOpacity>
+          <TouchableOpacity
+            style={[styles.fullButton, styles.kakaoButton]}
+            onPress={handleKakaoLogin}
+            activeOpacity={0.85}
+            disabled={isSocialLoading}
+          >
+            <KakaoIcon width={18} height={18} />
+            <Text style={[styles.fullButtonText, styles.kakaoText]}>
+              카카오 로그인
+            </Text>
+          </TouchableOpacity>
 
-            {/* 하단 링크 */}
-            <View style={styles.linkRow}>
-              <Text style={styles.linkText}>비밀번호 찾기</Text>
+          <TouchableOpacity
+            style={[styles.fullButton, styles.naverButton]}
+            onPress={handleNaverLogin}
+            activeOpacity={0.85}
+            disabled={isSocialLoading}
+          >
+            <NaverIcon width={16} height={16} />
+            <Text style={[styles.fullButtonText, styles.naverText]}>
+              네이버 로그인
+            </Text>
+          </TouchableOpacity>
 
-              <View style={{flex: 1}} />
-              <TouchableOpacity
-                onPress={() => navigation.navigate("NativeSignup")}
-                activeOpacity={0.7}
-              >
-                <Text style={styles.linkText}>회원가입</Text>
-              </TouchableOpacity>
-            </View>
-          </View>
-
-          <View>
-            {/* SNS 간편 로그인 구분선 */}
-            <View style={styles.snsSection}>
-              <View style={styles.line} />
-              <Text style={styles.snsLabel}>SNS 간편 로그인</Text>
-              <View style={styles.line} />
-            </View>
-
-            {/* SNS 버튼들 */}
-            <View style={styles.snsButtonsRow}>
-              <TouchableOpacity
-                style={[styles.snsButton, styles.kakaoButton]}
-                onPress={handleKakaoLogin}
-                activeOpacity={0.8}
-              >
-                <KakaoIcon width={50} aspectRatio={1} />
-              </TouchableOpacity>
-
-              <TouchableOpacity
-                style={[styles.snsButton, styles.naverButton]}
-                onPress={handleNaverLogin}
-                activeOpacity={0.8}
-              >
-                <NaverIcon width={50} aspectRatio={1} />
-              </TouchableOpacity>
-
-              <TouchableOpacity onPress={hardResetKakao}>
-                <Text style={{color: "white"}}>카카오 세션 초기화</Text>
-              </TouchableOpacity>
-            </View>
-          </View>
-        </KeyboardAvoidingView>
-      </TouchableWithoutFeedback>
+          {/* <TouchableOpacity onPress={hardResetKakao}>
+            <Text style={{color: "white"}}>카카오 세션 초기화</Text>
+          </TouchableOpacity> */}
+        </View>
+      </View>
     </SafeAreaView>
   );
 };
@@ -238,95 +165,61 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: "#000000",
   },
-  inner: {
+  container: {
     flex: 1,
-    paddingHorizontal: "4%",
-    paddingTop: "32%",
-    paddingBottom: 40,
-    paddingBottom: "32%",
-    justifyContent: "space-between",
+    paddingHorizontal: 20,
   },
-  logoWrapper: {
-    marginTop: 40,
-    alignItems: "center",
-  },
-  form: {
-    marginTop: 40,
-  },
-  input: {
-    height: 52,
-    borderRadius: 8,
-    paddingHorizontal: 16,
-    borderWidth: 1,
-    borderColor: "rgba(255,255,255,0.25)",
-    backgroundColor: "rgba(0,0,0,0.15)",
-    color: "#FFFFFF",
-    fontSize: 14,
-  },
-  inputSpacing: {
-    marginTop: 12,
-  },
-  loginButton: {
-    marginTop: 16,
-    height: 45,
-    borderRadius: 8,
-    backgroundColor: "#FFFFFF",
+
+  logoArea: {
+    flex: 1,
     alignItems: "center",
     justifyContent: "center",
   },
-  loginButtonText: {
-    fontSize: 16,
-    fontWeight: "700",
+
+  bottomArea: {
+    paddingBottom: 24,
+    gap: 12,
+  },
+
+  fullButton: {
+    height: 54,
+    borderRadius: 12,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    gap: 8,
+  },
+  fullButtonText: {
+    fontFamily: "NotoSansKR_SemiBold",
+    fontSize: 20,
+    // borderWidth: 1,
+    lineHeight: 27,
+  },
+
+  appleButton: {
+    backgroundColor: "#FFFFFF",
+  },
+  appleText: {
     color: "#111111",
   },
-  linkRow: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginTop: 12,
+  appleIcon: {
+    color: "#111111",
+    fontSize: 18,
+    fontWeight: "900",
+    marginTop: -1,
   },
-  linkText: {
-    fontSize: 14,
-    color: "#E0E0EA",
-  },
-  dividerDot: {
-    fontSize: 12,
-    color: "#E0E0EA",
-  },
-  snsSection: {
-    alignItems: "center",
-    flexDirection: "row",
-    marginTop: 40,
-    marginBottom: 16,
-  },
-  line: {
-    flex: 1,
-    height: StyleSheet.hairlineWidth,
-    backgroundColor: "rgba(255,255,255,0.4)",
-  },
-  snsLabel: {
-    marginHorizontal: 12,
-    fontSize: 12,
-    color: "#E0E0EA",
-    fontFamily: "NotoSansKR_SemiBold",
-  },
-  snsButtonsRow: {
-    flexDirection: "row",
-    justifyContent: "center",
-    gap: "7%",
-  },
-  snsButton: {
-    // width: 56,
-    // height: 56,
-    // borderRadius: 28,
-    alignItems: "center",
-    justifyContent: "center",
-    // borderWidth:1,
-    // borderColor:'white'
-  },
+
   kakaoButton: {
-    // backgroundColor: "#FEE500",
+    backgroundColor: "#FEE500",
   },
+  kakaoText: {
+    color: "#111111",
+  },
+
   naverButton: {
-    // backgroundColor: "#03C75A",
+    backgroundColor: "#03C75A",
+  },
+  naverText: {
+    color: "#FFFFFF",
   },
 });
